Replace a Pod =end regex with 'nextgroup'
This replaces the inefficient match for the Pod type name after =end,
which was attempted at the start of every line. Instead, we use
'nextgroup' after every delimited Pod region.

When highlighting t_source/perl6/basic.t, this pattern was the 2nd most
costly in terms of time spent, despite it never even successfully
matching in that file, so this is a nice performance boost.